requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
weasyprint>=66.0
pyyaml==6.0.1
jinja2==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.25.2,<0.26.0
ollama>=0.1.0
beautifulsoup4>=4.12.0
duckduckgo-search>=4.0.0
streamlit>=1.28.0

